requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
httpx>=0.25.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
psutil>=5.9.0
pywinauto>=0.6.8
